requirements.txt
Django
django-apscheduler
imapclient
django-ckeditor